second batch of moving files around
- moved everything to be moved into `./supibot` including type files
- also marked several singletons to be ported over
- all definitions adjusted for the split
- removed some usages of supibot-specific modules from generic modules
