+++ @types/index.d.ts
@@ -9,14 +9,9 @@ import { GotProxy as Got } from "./classes/got";
 
 // singletons
 import { CacheSingleton } from "./singletons/cache";
-import { CooldownManagerSingleton } from "./singletons/cooldown-manager";
-import { LoggerSingleton } from "./singletons/logger";
 import { MetricsSingleton } from "./singletons/metrics";
-import { PastebinSingleton } from "./singletons/pastebin";
 import { QuerySingleton } from "./singletons/query";
-import { SandboxSingleton } from "./singletons/sandbox";
 import { UtilsSingleton } from "./singletons/utils";
-import { VLCSingleton } from "./singletons/vlc-connector";
 
 export declare type GlobalSbObject = {
     Date: typeof CustomDate,
@@ -27,14 +22,9 @@ export declare type GlobalSbObject = {
     Got: typeof Got,
 
     Cache: InstanceType<typeof CacheSingleton>,
-    CooldownManager: InstanceType<typeof CooldownManagerSingleton>,
-    Logger: InstanceType<typeof LoggerSingleton>,
     Metrics: InstanceType<typeof MetricsSingleton>,
-    Pastebin: InstanceType<typeof PastebinSingleton>,
     Query: InstanceType<typeof QuerySingleton>,
-    Sandbox: InstanceType<typeof SandboxSingleton>,
     Utils: InstanceType<typeof UtilsSingleton>,
-    VideoLANConnector: InstanceType<typeof VLCSingleton>
 };
 
 // declare type ModuleName = keyof GlobalSbObject;
@@ -46,16 +36,10 @@ declare type ModuleFilePath = "classes/afk"
     | "objects/errors"
     | "objects/promise"
     | "singletons/cache"
-    | "singletons/cooldown-manager"
-    | "singletons/logger"
     | "singletons/metrics"
-    | "singletons/pastebin"
     | "singletons/query"
-    | "singletons/sandbox"
-    | "singletons/system-log"
     | "singletons/twitter"
     | "singletons/utils"
-    | "singletons/vlc-connector";
 
 declare type OptionsObject = {
     blacklist?: ModuleFilePath[];
