+++ supibot/classes/afk-responses.json
