+++ supibot/classes/afk.js
