+++ supibot/classes/banphrase.js
