+++ supibot/classes/channel.js
