+++ supibot/classes/chat-module.js
