+++ supibot/classes/command.js
