+++ supibot/classes/filter.js
