+++ supibot/classes/platform.js
