+++ supibot/classes/template.js
