+++ supibot/classes/user.js
