+++ supibot/singletons/vlc-client.js
