+++ supibot/singletons/vlc-connector.js
@@ -28,7 +28,7 @@ const mandatoryConfigs = [
 /**
  * VideoLANConnector (VLC) handler module - handles a VLC instance's playlist and methods.
  */
-module.exports = class VLCSingleton extends require("./template.js") {
+module.exports = class VLCSingleton extends require("../../singletons/template.js") {
 	/**
 	 * @inheritDoc
 	 * @returns {VLCSingleton}
@@ -296,47 +296,6 @@ module.exports = class VLCSingleton extends require("./template.js") {
 		}
 		else {
 			return status.information;
-		}
-	}
-
-	/**
-	 * Deletes the last video queued by a certain user.
-	 * Used when the user queues a song they didn't want to queue.
-	 * Returns void if the user was not found or that user has no active requests.
-	 * Returns song name if the deletion was successful.
-	 * @param {number} user ID
-	 * @returns {Promise<void|string>}
-	 */
-	async wrongSong (user) {
-		const userData = await sb.User.get(user);
-		if (!userData) {
-			return { success: false, reason: "no-user" };
-		}
-
-		const userRequests = this.videoQueue.filter(i => i.user === userData.ID);
-		if (userRequests.length === 0) {
-			return { success: false, reason: "no-requests" };
-		}
-
-		const playingData = await this.currentlyPlayingData();
-		if (!playingData || !Number(playingData.vlcID)) {
-			playingData.vlcID = -Infinity;
-		}
-
-		try {
-			const [firstUserSong] = userRequests
-				.filter(i => i.vlcID >= playingData.vlcID)
-				.sort((a, b) => a.vlcID - b.vlcID);
-
-			const link = sb.Utils.linkParser.parseLink(firstUserSong.link);
-			const deletedSongData = await this.getDataByName(firstUserSong.name, link);
-			await this.delete(deletedSongData.id);
-
-			return { success: true, song: deletedSongData };
-		}
-		catch (e) {
-			console.error(e);
-			return { success: false, reason: "delete-failed" };
 		}
 	}
